Create Q_1_A.py
Solution for question 1_A with error testing and code with baked error testing and redundancy.
This code is purposely written with some issues (it runs) so you can try seeing to improve it.

If you have an improvement, fork the repo and create your own branch, then create a pull request.
In later solutions, there may be extra credit attached for this.